Patch 9 into koen84 master

+++ script.sh
@@ -15,22 +15,25 @@ case "$1" in
 
 'install')
   read -p "How many nodes do you want to run ? : " NUMBEROFNODES
-  if [ $NUMBEROFNODES = "" ]
+  re='^[0-9]+$'
+  if ! [[ $NUMBEROFNODES =~ $re ]] && [ "$NUMBEROFNODES" -gt 0 ]
   then
       NUMBEROFNODES = 1
   fi
   
-  prerequisites
-  replicant
   #Check if CUSTOM_HOME exists
   if ! [ -d "$CUSTOM_HOME" ]; then echo -e "${RED}Please configure your variables first ! (variables.cfg --> CUSTOM_HOME & CUSTOM_USER)${NC}"; exit; fi
   
+  install_sudofix1
+  prerequisites
+  replicant
+
   #Keep track of how many nodes you've started on the machine
   echo "$NUMBEROFNODES" > $CUSTOM_HOME/.numberofnodes
   paths
   go_lang
   #If repos are present and you run install again this will clean up for you :D
-  if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then echo -e "${RED}--> Repos present. Please run update.sh script...${NC}"; echo -e; exit; fi
+  if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then echo -e "${RED}--> Repos present. Either run the upgrade command or cleanup & install again...${NC}"; echo -e; exit; fi
   mkdir -p $GOPATH/src/github.com/ElrondNetwork
   git_clone
   build_node
@@ -42,6 +45,7 @@ case "$1" in
          INDEX=$(( $i - 1 ))
          WORKDIR="$CUSTOM_HOME/elrond-nodes/node-$INDEX"
          install
+         install_utils
          node_name
          keys
          systemd
@@ -68,6 +72,7 @@ case "$1" in
            echo -e "${GREEN}I'll take that as a no then...${NC}"
             ;;
       esac
+  sudo chown -R $CUSTOM_USER:$CUSTOM_USER $CUSTOM_HOME/elrond-nodes
   ;;
 
 'install-remote')
@@ -91,6 +96,9 @@ case "$1" in
   if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then sudo rm -rf $GOPATH/src/github.com/ElrondNetwork/elrond-*; echo -e; echo -e "${RED}--> Repos present. Removing and fetching again...${NC}"; echo -e; fi
   git_clone
   build_node
+  if ! [ -d "$CUSTOM_HOME/elrond-utils" ]; then mkdir -p $CUSTOM_HOME/elrond-utils; fi
+  
+  install_utils
   
   INSTALLEDNODES=$(cat $CUSTOM_HOME/.numberofnodes)
   
@@ -121,9 +129,10 @@ case "$1" in
 'auto_upgrade')
   paths
   #Remove previously cloned repos
-  if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then sudo rm -rf $GOPATH/src/github.com/ElrondNetwork/elrond-*; fi
+  if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then rm -rf $GOPATH/src/github.com/ElrondNetwork/elrond-*; fi
   git_clone
   build_node
+  install_utils
   
   INSTALLEDNODES=$(cat $CUSTOM_HOME/.numberofnodes)  
   curl --silent "https://api.github.com/repos/ElrondNetwork/elrond-go/releases/latest" | grep "body" > $HOME/tmp
@@ -145,7 +154,7 @@ if [ "$DBQUERY" -eq "1" ]; then
                         cleanup
                         update
                         mv $UPDATEWORKDIR/config/prefs.toml.save $UPDATEWORKDIR/config/prefs.toml
-                        sudo systemctl start elrond-node-$UPDATEINDEX       
+                        sudo systemctl start elrond-node-$UPDATEINDEX
                       done
       
     else
@@ -233,30 +242,53 @@ if [ "$DBQUERY" -eq "1" ]; then
        [Yy]* )
           echo -e "${RED}OK ! Cleaning everything !${NC}"
           
-          NODESTODESTROY=$(cat $CUSTOM_HOME/.numberofnodes)
-              for i in $(seq 1 $NODESTODESTROY);
-                  do
-                      KILLINDEX=$(( $i - 1 ))
-                        echo -e
-                        echo -e "${GREEN}Stopping Elrond Node-$KILLINDEX binary on host ${CYAN}$HOST${GREEN}...${NC}"
-                        echo -e
-                        [ -e /etc/systemd/system/elrond-node-$KILLINDEX.service ] && sudo systemctl stop elrond-node-$KILLINDEX
-                        echo -e "${GREEN}Erasing unit file and node folder for Elrond Node-$KILLINDEX...${NC}"
-                        echo -e
-                        [ -e /etc/systemd/system/elrond-node-$KILLINDEX.service ] && sudo rm /etc/systemd/system/elrond-node-$KILLINDEX.service
-                        if [ -d $CUSTOM_HOME/elrond-nodes/node-$KILLINDEX ]; then sudo rm -rf $CUSTOM_HOME/elrond-nodes/node-$KILLINDEX; fi
-                  done
+          if [[ -f $CUSTOM_HOME/.numberofnodes ]]; then
+            NODESTODESTROY=$(cat $CUSTOM_HOME/.numberofnodes)
+                for i in $(seq 1 $NODESTODESTROY);
+                    do
+                        KILLINDEX=$(( $i - 1 ))
+                          echo -e
+                          echo -e "${GREEN}Stopping Elrond Node-$KILLINDEX binary on host ${CYAN}$HOST${GREEN}...${NC}"
+                          echo -e
+                          if [ -e /etc/systemd/system/elrond-node-$KILLINDEX.service ]; then sudo systemctl stop elrond-node-$KILLINDEX; fi
+                          echo -e "${GREEN}Erasing unit file and node folder for Elrond Node-$KILLINDEX...${NC}"
+                          echo -e
+                          if [ -e /etc/systemd/system/elrond-node-$KILLINDEX.service ]; then sudo rm /etc/systemd/system/elrond-node-$KILLINDEX.service; fi
+                          if [ -d $CUSTOM_HOME/elrond-nodes/node-$KILLINDEX ]; then sudo rm -rf $CUSTOM_HOME/elrond-nodes/node-$KILLINDEX; fi
+                    done
+          fi
             
             #Reload systemd after deleting node units
             sudo systemctl daemon-reload
+            
+            echo -e
+            echo -e "${GREEN}Removing elrond utils...${NC}"
+            echo -e      
+            
+            if ps -all | grep -q termui; then killall termui; sleep 2; fi
+            if [[ -e $CUSTOM_HOME/elrond-utils/termui ]]; then rm $CUSTOM_HOME/elrond-utils/termui; fi
+              
+            if ps -all | grep -q logviewer; then killall logviewer; sleep 2; fi
+            if [[ -e $CUSTOM_HOME/elrond-utils/logviewer ]]; then rm $CUSTOM_HOME/elrond-utils/logviewer; fi
+            
+            rm -rf $CUSTOM_HOME/elrond-utils && rm -rf $CUSTOM_HOME/elrond-nodes
+            if [[ -e $CUSTOM_HOME/autoupdate.status ]]; then rm $CUSTOM_HOME/autoupdate.status; fi 
+            
             echo -e
             echo -e "${GREEN}Removing auto-updater crontab from host ${CYAN}$HOST${GREEN}...${NC}"
             echo -e      
-            crontab -l | grep -v 'elrond-go-scripts-v2/auto-updater.sh'  | crontab -
+            crontab -l | grep -v '/auto-updater.sh'  | crontab -
+            
+            echo -e "${GREEN}Removing paths from .profile on host ${CYAN}$HOST${GREEN}...${NC}"
+            echo -e
+            sed -i 'N;$!P;$!D;$d' ~/.profile
             
             echo -e "${GREEN}Removing cloned elrond-go & elrond-configs repo from host ${CYAN}$HOST${GREEN}...${NC}"
             echo -e      
             if [ -d "$GOPATH/src/github.com/ElrondNetwork/elrond-go" ]; then sudo rm -rf $GOPATH/src/github.com/ElrondNetwork/elrond-*; fi      
+            
+            # remove the sudofix file
+            sudo rm -f $sudofix
             ;;
             
        [Nn]* )
@@ -292,7 +324,7 @@ if [ "$DBQUERY" -eq "1" ]; then
   case $yn in
        [Yy]* )
           echo -e "${GREEN}Adding auto-update to crontab !${NC}"
-          if (crontab -l | grep -q "auto-updater.sh"); then echo "Crontab already installed"; else autoupdate; fi  
+          if (crontab -l 2>/dev/null | grep -q "auto-updater.sh"); then echo "Crontab already installed"; else autoupdate; fi  
             ;;
        [Nn]* )
           echo -e "${GREEN}Fine... let's continue...${NC}"
@@ -322,6 +354,8 @@ if [ "$DBQUERY" -eq "1" ]; then
   ;;
   
 'sudofix')
+        install_sudofix1
+
         INSTALLEDNODES=$(cat $CUSTOM_HOME/.numberofnodes)
         for i in $(seq 1 $INSTALLEDNODES);
                 do
